pyproject.toml
[project]
name = "llm-training-stack"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.0.0",
  "transformers>=4.40.0",
  "datasets>=2.18.0",
  "peft>=0.11.0",
  "deepspeed>=0.14.0",
  "prometheus-client>=0.20.0",
  "psutil>=5.9.0",
  "pynvml>=11.5.0",
  "pyyaml>=6.0.0",
  "hydra-core>=1.3.2",
  "loguru>=0.7.2",
]

[project.scripts]
train-llm = "training.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["training*", "serving*"]

[tool.setuptools.package-data]
"training.configs" = ["*.yaml"]
